% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calculate_hw_grade}
\alias{calculate_hw_grade}
\title{Calculate Homework Grades in the style of Compass2g (Blackboard)}
\usage{
calculate_hw_grade(hw_data, hw_name = "hw01")
}
\arguments{
\item{hw_data}{Either a `data.frame` or `tibble` containing multiple variables
with the prefix `ex` that are numeric and `netid`.}

\item{hw_name}{The name of assessment. Default is `"hw01"`.}
}
\value{
A `tibble` with `Username` and `hw_name` variables.
}
\description{
Summarize multiple exercise columns into one and calculate output
}
\examples{
\dontrun{
calculate_hw_grade(hw)

calculate_hw_grade(hw, hw_name = "hw03")
}
}
