% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coursetools-package.R
\docType{package}
\name{coursetools-package}
\alias{coursetools}
\alias{coursetools-package}
\title{coursetools: Automates Course Work}
\description{
Provides a pathway to orderly establish grades, grant and revoke
git permissions and more.
}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu}

}
