% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roster.R
\name{extract_student_ids}
\alias{extract_student_ids}
\title{Extract Student IDs from Mailing List}
\usage{
extract_student_ids(emails, domain = "illinois.edu", delim = ";")
}
\arguments{
\item{emails}{A list of e-mails with format:
`netid1@domain.edu;netid2@domain.edu;...`}

\item{domain}{The domain of the e-mail address, e.g. `domain.edu`
or `illinois.edu`.}

\item{delim}{Value that is separating student e-mail address, e.g. `;` `,` `\\n`.}
}
\value{
A `character vector` containing student IDs.
}
\description{
Takes a mailing list form and extracts the student's ID from it.
}
\examples{
registered_students = "netid1@domain.edu;netid2@domain.edu"

extract_student_ids(registered_students, domain = "domain.edu")

registered_students_comma = "netid1@domain.edu,netid2@domain.edu"
extract_student_ids(registered_students_comma, domain = "domain.edu", delim = ",")
}
