% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveCode.R
\name{r_to_html}
\alias{r_to_html}
\title{Renders r file directly to html via Quarto}
\usage{
r_to_html(r_file, author)
}
\arguments{
\item{r_file}{Path to R file}
\item{author}{Author name to be included Quarto YAML header}
}
\value{
Outputs 1 html file
}
\description{
Takes r file, generates QMD file, and renders to html; returning only html. 
}
