% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{read_directory}
\alias{read_directory}
\title{Read Multiple Flat Files Into _R_ Simultaneously}
\usage{
read_directory(dir, pattern = NULL, delim = ",")
}
\arguments{
\item{dir}{Location of the files to be read into R.}

\item{pattern}{The file extension or pattern to match. Default `NULL` returns
all files in a directory.}

\item{delim}{The delimiter between successive observations. Default `,`.}
}
\value{
A `tibble` containing all files read into _R_.
}
\description{
Searches for flat files of a specific type and reads them into _R_ to form
one large `tibble`
}
\examples{
\dontrun{
# Relative path
hw03_grades = read_directory("grades/hw/hw03-grades")

# Only CSVs
hw01_grades = read_directory("grades/hw/hw01-grades", pattern = ".csv", delim = ",")
}
}
