% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveCode.R
\name{saveCode}
\alias{saveCode}
\title{Saves all code chunks as R and txt files}
\usage{
saveCode(last_name)
}
\arguments{
\item{last_name}{User's last name}
}
\value{
Outputs 1 R file and 1 txt file
}
\description{
Saves all code chunks as R and txt files
}
