% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{write_blackboard_hw_grades}
\alias{write_blackboard_hw_grades}
\title{Write Homework Grades in the style of Compass2g (Blackboard)}
\usage{
write_blackboard_hw_grades(grades, hw_name = "hw01", output_dir = ".")
}
\arguments{
\item{grades}{Location of the files to be read into R.}

\item{pattern}{The file extension or pattern to match. Default `NULL` returns
all files in a directory.}

\item{delim}{The delimiter between successive observations. Default `,`.}
}
\value{
A `tibble` containing grades and a CSV on file system.
}
\description{
Generate a Comma-Separated-Values (CSV) file containing the grades for one
homework assignment.
}
\examples{
\dontrun{
write_blackboard_hw_grades(grades)
write_blackboard_hw_grades(grades, hw_name = "hw03")
write_blackboard_hw_grades(grades, hw_name = "hw02", output_dir = "hw/hw02")
}
}
